requirements.txt
flask
SQLAlchemy
databases[postgresql]
psycopg2-binary
jsonify
werkzeug